Initial release (#1)
* Initial project structure except Actions

This project is going to be TDD since it's very well scoped. I'm 98%
sure this will be all tied up with a pretty bow in a few days and will
not really ever need to be modified

* Add placeholder environment variable scripts

This is just so they don't get missed as they are referenced at the
bottom of the readme

* Fix typo in readme

* Add placeholders for other referenced sections in readme

* Improve API names, add docstrings

This renames things to have their exact functions be as clear as
possible, and removes boolean arguments. Furthermore this
improves general consistency. Example changes:

- coll.ensure() became coll.create_if_not_exists()
- coll.read(key) became coll.read_doc(key)
- doc.replace() became doc.compare_and_swap()
- doc.read(try_304=True) became doc.read_if_remote_newer()

* More work simplifying API

* Update readme JWT section

* Explicitly document where JWTAuth gets wrapped

* Fix Readme example deleting doc it's still using

* Fix dead code in example

* Simplify readme simple caching example

* More documentation/scaffolding for classes

Next up is documenting "env_config" and then should be
ready to start tests

* Document env_config signature, add timeouts.

Also noted that it's easier for testing and nicer for compatibility
if env_config allows passing a dict instead of always using an
environment variable. Furthermore it doesn't require any extra
work and only has slightly less clear error messages.

* Improve readme formatting

* Document username/password in env_config

* Improve JWT recommendation in env_config

* Clarify common deviations for jwt in env_config

* Fix type in env_config docstring

* Get through a good chunk of tests

* Finish initial test suite

* Fix using instance method as class method

* Stub additional examples

* Implement env_config

* Implement basic auth

This means that we now have the main logic for "how to send a
request to Arango" complete!

* Work on JWT auth

* Finish JWT Auth

The boolean argument to the helper http methods isn't
great but doubling the number of methods isn't great either

* Implement to a working test suite

82% coverage; not bad at all for a first pass

* Fix flake8 warnings

* Finish examples

* Ensure examples have no unused code

* Improve coverage report

I'm pretty happy with 85%; most of the missing cases are
errors

* Add initial workflow

* Fix workflow env vars

* Fix port issue

* Try with root password

* Fix wrong envvars in examples

* Show coverage missing lines in examples

* Upload to pypi