Fix JWT corruption not automatically recovering on first request (#5)
This fixes JWT corruption and expiration not recovering on the first request if they were being loaded from a disk cache, since although it properly cleared the token the first time around, it just immediately loaded the exact same token. It will now detect if it just got the exact same token from the cache and if so it will disregard it.

This will require at most one refresh per token, even in parallel contexts, when the disk cache corrupted.